feat: add more-info page
The page shows the info about the picture of the day (although just a
placeholder text is used right now). This is different from the index
page which only shows the pircure of the day and the date.

An error image is shown in case there are fetch errors.

The page is not styled, yet.